+++ src/fuzzy_search_levenshtein.rs
@@ -99,13 +99,12 @@ impl<'a> Iterator for FuzzySearchLevenshtein<'a> {
                     if candidate.text_index <= self.text.len() {
                         if candidate.distance == 0 {
                             self.candidates.clear();
+                            self.current_text_index += 1;
+                            self.candidates.push(CandidateMatch::new(
+                                self.current_text_index,
+                                self.current_text_index,
+                            ));
                         }
-
-                        self.current_text_index += 1;
-                        self.candidates.push(CandidateMatch::new(
-                            self.current_text_index,
-                            self.current_text_index,
-                        ));
 
                         self.best_found_distance = candidate.distance;
                         return Some(candidate);
@@ -136,5 +135,40 @@ impl<'a> Iterator for FuzzySearchLevenshtein<'a> {
         }
 
         None
+    }
+}
+
+#[cfg(test)]
+mod fuzzy_search_levenshtein_tests {
+    use crate::{
+        fuzzy_search_levenshtein::FuzzySearchLevenshtein, fuzzy_search_options::FuzzySearchOptions,
+    };
+
+    #[test]
+    fn test_all_results() {
+        run_find_levenshtein_all(
+            "pattern",
+            "---------------------pattttern",
+            // 21,
+            // "pattttern",
+            // 2,
+            3,
+        );
+    }
+
+    fn run_find_levenshtein_all(
+        pattern: &str,
+        text: &str,
+        // expected_start: usize,
+        // expected_match: &str,
+        // expected_distance: usize,
+        max_distance: usize,
+    ) {
+        let options = FuzzySearchOptions::new(max_distance);
+        let all_results = FuzzySearchLevenshtein::find(pattern, text, &options).collect::<Vec<_>>();
+
+        println!("{:?}", all_results);
+
+        assert_eq!(all_results.len(), 24);
     }
 }
